requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
mingus==0.6.1
python-multipart==0.0.20
loguru==0.7.3
pylint==3.3.7
black==25.1.0
openai==1.30.1
pretty_midi==0.2.10
setuptools
python-dotenv==1.0.1
